Added UC-/Charter-relevant e-mails
I trolled through the SACM email I have to find messages with content
relevant to our current use case and charter efforts.  These are in PDF
form, and the file names try to distinguish the subject matter to a
(limited) degree.

There really aren't many notes in the way of specific guidance at the
functional capability/component level, but many provide higher level
guidance with respect to overall document content and/or structure.